Fix as to ignore --save-temps
mips/gcc4.6 is now built with binutils-2.24 to be consistant
with gcc4.8 and gcc4.9 on nan handling.  Unfortunately
gcc4.6 pass on -save-temps to not recognized by "as".
gcc4.6 is normally built with binutils-2.21 which has fix for
as to accept but ignore -save-temps.  This CL catptures the
same fix in binutils-2.21

Change-Id: I58095a67810f2bcf6e7594383002247a9491a901